chore: 파일 분리 및 로그 제거
