+++ FoodDiary/Data/Sources/Repository/FoodRecordRepositoryImpl.swift
@@ -8,40 +8,6 @@ import Foundation
 import Photos
 import UIKit
 
-// MARK: - In-memory LRU Cache
-
-private actor PhotoURLCache {
-    private var entries: [(key: String, value: [Date: [URL]])] = []
-    private let capacity = 12
-
-    func get(for dateRange: ClosedRange<Date>) -> [Date: [URL]]? {
-        let k = cacheKey(for: dateRange)
-        guard let index = entries.firstIndex(where: { $0.key == k }) else { return nil }
-        let entry = entries.remove(at: index)
-        entries.insert(entry, at: 0)
-        return entry.value
-    }
-
-    func set(_ value: [Date: [URL]], for dateRange: ClosedRange<Date>) {
-        let k = cacheKey(for: dateRange)
-        entries.removeAll { $0.key == k }
-        entries.insert((key: k, value: value), at: 0)
-        if entries.count > capacity {
-            let evicted = entries.removeLast()
-            print("[PhotoURLCache] 캐시 용량 초과 — 삭제: \(evicted.key)")
-        }
-    }
-
-    func remove(for dateRange: ClosedRange<Date>) {
-        let k = cacheKey(for: dateRange)
-        entries.removeAll { $0.key == k }
-    }
-
-    private func cacheKey(for dateRange: ClosedRange<Date>) -> String {
-        "\(Int(dateRange.lowerBound.timeIntervalSince1970))-\(Int(dateRange.upperBound.timeIntervalSince1970))"
-    }
-}
-
 /// FoodRecordRepository 구현체
 public struct FoodRecordRepositoryImpl<
     Client: HTTPClienting & Sendable,
@@ -135,14 +101,9 @@ public struct FoodRecordRepositoryImpl<
     public func fetchPhotoURLs(in dateRange: ClosedRange<Date>) -> AsyncThrowingStream<[Date: [URL]], Error> {
         AsyncThrowingStream { continuation in
             Task {
-                let rangeLabel = "\(formatDate(dateRange.lowerBound)) ~ \(formatDate(dateRange.upperBound))"
-
                 // 1. 캐시 HIT 시 즉시 방출
                 if let cached = await photoURLCache.get(for: dateRange) {
-                    print("[FoodRecordRepository] Cache HIT: \(rangeLabel)")
                     continuation.yield(cached)
-                } else {
-                    print("[FoodRecordRepository] Cache MISS: \(rangeLabel) — fetching...")
                 }
 
                 // 2. 항상 서버 검증
@@ -173,10 +134,7 @@ public struct FoodRecordRepositoryImpl<
                     let current = await photoURLCache.get(for: dateRange)
                     if current != fresh {
                         await photoURLCache.set(fresh, for: dateRange)
-                        print("[FoodRecordRepository] Cache UPDATED: \(rangeLabel)")
                         continuation.yield(fresh)
-                    } else {
-                        print("[FoodRecordRepository] Cache VALID: \(rangeLabel)")
                     }
 
                     continuation.finish()
